ai_server_proto/requirements.txt
beautifulsoup4==4.6.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.2.1
django-debug-toolbar==1.11
django-filter==2.1.0
django-modelcluster==4.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.9.4
draftjs-exporter==2.1.6
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-storage==1.16.0
google-cloud-vision==0.36.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
gunicorn==19.9.0
html5lib==1.0.1
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==5.4.1
protobuf==3.8.0
psycopg2-binary==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
Pygments==2.4.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.25.3
wagtail==2.5.1
webencodings==0.5.1
Willow==1.1
